Abstract username form to component
